merkle tree module created
