+++ mithril-stm/src/error.rs
@@ -1,6 +1,7 @@
 //! Crate specific errors
 
-use crate::merkle_tree::{BatchPath, Path};
+use crate::merkle_tree::basic::Path;
+use crate::merkle_tree::batch_compatible::BatchPath;
 use blake2::digest::{Digest, FixedOutput};
 use {
     crate::multi_sig::{Signature, VerificationKey, VerificationKeyPoP},
