+++ mithril-stm/src/key_reg.rs
@@ -1,7 +1,8 @@
 //! Key registration functionality.
 use super::stm::Stake;
 use crate::error::RegisterError;
-use crate::merkle_tree::{MTLeaf, MerkleTree};
+use crate::merkle_tree::leaf::MTLeaf;
+use crate::merkle_tree::tree::MerkleTree;
 use crate::multi_sig::{VerificationKey, VerificationKeyPoP};
 use blake2::digest::{Digest, FixedOutput};
 use std::collections::hash_map::Entry;
