+++ mithril-stm/src/merkle_tree/mod.rs
@@ -1,0 +1,6 @@
+//! Merkle tree implementation for STM
+
+pub(super) mod basic;
+pub(super) mod batch_compatible;
+pub(super) mod leaf;
+pub(super) mod tree;
